add support to web: update restaurant module

+++ lib/app/modules/restaurant/restaurant_details/restaurant_details_controller.dart
@@ -1,8 +1,11 @@
+import 'dart:typed_data';
+
 import 'package:coffee_tracker/app/modules/restaurant/restaurant_controller.dart';
 import 'package:coffee_tracker/app/shared/auth/auth_controller.dart';
 import 'package:coffee_tracker/app/shared/models/restaurant_model.dart';
 import 'package:coffee_tracker/app/shared/repositories/storage/interfaces/media_storage_repository_interface.dart';
 import 'package:coffee_tracker/app/shared/repositories/storage/interfaces/storage_repository_interface.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:mobx/mobx.dart';
@@ -16,53 +19,39 @@ class RestaurantDetailsController = _RestaurantDetailsControllerBase
     with _$RestaurantDetailsController;
 
 abstract class _RestaurantDetailsControllerBase with Store {
-  _RestaurantDetailsControllerBase() {
-    final RestaurantModel r = Modular.args.data;
-    mediaStorage = Modular.get();
-    mediaStorage.loadCache();
-
-    setRestaurant(r);
-  }
-
   IMediaStorageRepository mediaStorage;
 
   @observable
   RestaurantModel restaurant;
 
   @observable
-  Image savedImage;
+  bool favorite;
 
-  @action
-  void setRestaurant(RestaurantModel value) {
-    restaurant = value;
-    loadImage('${restaurant.id}.jpg');
+  @observable
+  Uint8List savedImage;
+
+  _RestaurantDetailsControllerBase() {
+    final RestaurantModel r = Modular.args.data;
+    mediaStorage = Modular.get();
+
+    setRestaurant(r);
   }
 
-  @action
-  void setFavorite(bool value) {
-    restaurant.favorite = value;
+  Map<String, String> get shareData {
+    final AuthController auth = Modular.get();
+    final name = auth.user.displayName ?? 'User';
 
-    final IStorageRepository storage = Modular.get();
-
-    storage.updateRestaurant(restaurant);
-
-    loadImage('${restaurant.id}.jpg');
-  }
-
-  @action
-  void edit() {
-    Modular.to.pushNamed('restaurant/edit', arguments: restaurant);
-  }
-
-  @action
-  void showFullImage() {
-    Modular.to.showDialog(child: FullScreenImage(savedImage));
+    return {
+      'text': restaurant.getRecommendationText(name),
+      'subject': 'Recomendação de Café: ${restaurant.name}',
+    };
   }
 
   @action
   Future<void> delete() async {
-    mediaStorage.deleteRestaurantImage(
-        photoId: restaurant.imageURL, restaurantId: restaurant.id);
+    if (!kIsWeb)
+      mediaStorage.deleteRestaurantImage(
+          photoId: restaurant.imageURL, restaurantId: restaurant.id);
 
     final IStorageRepository storage = Modular.get();
     await storage.deleteRestaurant(restaurant.id);
@@ -73,39 +62,58 @@ abstract class _RestaurantDetailsControllerBase with Store {
     Modular.navigator.pop();
   }
 
-  @action
-  Future<void> setImage(Image value) async {
-    savedImage = value;
+  void edit() {
+    Modular.link.pushNamed('/edit', arguments: restaurant);
   }
 
-  Map<String, String> get shareData {
-    final path = restaurant.imageURL.isEmpty
-        ? ''
-        : mediaStorage.getFullPath(restaurant.imageURL);
+  void goToHome() {
+    Modular.to.popUntil(ModalRoute.withName('/home'));
+  }
 
-    final AuthController auth = Modular.get();
-    final name = auth.user.displayName;
-
-    return {
-      'text': restaurant.getRecommendationText(name),
-      'subject': 'Recomendação de Café: ${restaurant.name}',
-      'path': path,
-    };
+  void closePage() {
+    Modular.to.pop();
   }
 
   Future<void> loadImage(String fileName) async {
     if (restaurant.imageURL.isEmpty) {
-      setImage(Image.asset('images/no-image.png'));
+      setImage(null);
     } else {
-      setImage(await mediaStorage.fetchRestaurantImage(
-        restaurantId: restaurant.id,
-        photoId: restaurant.imageURL,
-      ));
+      setImage(
+        await mediaStorage.fetchRestaurantImage(
+          restaurantId: restaurant.id,
+          photoId: restaurant.imageURL,
+        ),
+      );
     }
   }
 
   Future<void> reload() async {
-    await mediaStorage.loadCache();
+    mediaStorage.loadCache();
     await loadImage(restaurant.imageURL);
+  }
+
+  @action
+  void setFavorite(bool value) {
+    favorite = value;
+    restaurant.favorite = value;
+    Modular.get<IStorageRepository>().updateRestaurant(restaurant);
+  }
+
+  @action
+  Future<void> setImage(Uint8List value) async {
+    savedImage = value != null && value.isNotEmpty ? value : null;
+  }
+
+  @action
+  void setRestaurant(RestaurantModel value) {
+    restaurant = value;
+    favorite = restaurant.favorite;
+    loadImage('${restaurant.id}.jpg');
+  }
+
+  @action
+  void showFullImage() {
+    if (savedImage != null)
+      Modular.to.showDialog(child: FullScreenImage(savedImage));
   }
 }
